+++ lib/src/ast/value.dart
@@ -5,11 +5,11 @@
 library toml.src.ast.value;
 
 import 'package:petitparser/petitparser.dart';
-import 'package:toml/encoder.dart';
 import 'package:toml/src/decoder/parser/util/non_strict.dart';
+import 'package:toml/src/encoder.dart';
 
-import 'node.dart';
 import 'key.dart';
+import 'node.dart';
 import 'value/array.dart';
 import 'value/boolean.dart';
 import 'value/datetime.dart';