+++ lib/src/decoder.dart
@@ -1,0 +1,10 @@
+// Copyright (c) 2015 Justin Andresen. All rights reserved.
+// This software may be modified and distributed under the terms
+// of the MIT license. See the LICENSE file for details.
+
+library toml.src.decoder;
+
+export 'decoder/exception/invalid_escape_sequence.dart';
+export 'decoder/exception/not_a_table.dart';
+export 'decoder/exception/redefinition.dart';
+export 'decoder/map_builder.dart';