+++ lib/src/decoder/map_builder.dart
@@ -4,7 +4,7 @@
 
 library toml.src.ast.decoder.map_builder;
 
-import 'package:toml/ast.dart';
+import 'package:toml/src/ast.dart';
 
 import 'exception/not_a_table.dart';
 import 'exception/redefinition.dart';