+++ lib/src/encoder.dart
@@ -1,0 +1,11 @@
+// Copyright (c) 2015 Justin Andresen. All rights reserved.
+// This software may be modified and distributed under the terms
+// of the MIT license. See the LICENSE file for details.
+
+library toml.src.encoder;
+
+export 'encoder/ast_builder.dart';
+export 'encoder/encodable.dart';
+export 'encoder/exception/mixed_array_types.dart';
+export 'encoder/exception/unknown_value_type.dart';
+export 'encoder/pretty_printer.dart';