+++ lib/src/encoder/exception/unknown_value_type.dart
@@ -4,7 +4,7 @@
 
 library toml.src.encoder.exception.unknown_value_type;
 
-import 'package:toml/exception.dart';
+import 'package:toml/src/exception.dart';
 
 /// An error which is thrown when an object cannot be encoded.
 ///