+++ lib/src/encoder/pretty_printer.dart
@@ -4,7 +4,7 @@
 
 library toml.src.encoder.pretty_printer;
 
-import 'package:toml/ast.dart';
+import 'package:toml/src/ast.dart';
 
 /// A visitor for TOML AST nodes that converts them to a TOML formatted string.
 ///