+++ test/tester/value.dart
@@ -5,7 +5,7 @@
 library toml.test.tester.value;
 
 import 'package:petitparser/petitparser.dart';
-import 'package:toml/ast.dart';
+import 'package:toml/toml.dart';
 
 import 'toml.dart';