+++ CHANGELOG.md
@@ -10,6 +10,9 @@ This is a major update that does not only bring along many changes to the intern
   Use `TomlDocument.fromMap` and `TomlDocument.toString` instead.
 - Removed the `toml.loader` library.
   Use `TomlDocument.load` instead.
+- Removed the `yaml` dependency.
+  JSON and YAML documents cannot be loaded with this package anymore.
+  Since YAML documents are needed for testing, there is still a `dep_dependency` for the `yaml` package.
 - In preparation to support TOML v0.5.0 in an upcoming version of the library, the parser is now based on TOML's official ABNF.
 
 ## 0.6.1 / 2020-11-19